requirements.txt
requests~=2.18
beautifulsoup4~=4.6
ruamel.yaml<0.15
